Improves RTS queue to be dynamic and VC specific
For fault tolerance use, a RTS queue is added in [81b3911a] to track shm
LMT RTS messages. However, the queue is global and static, which may not
be scalable.

This patch moves the RTS queue to struct MPIDI_CH3I_VC, to be VC
specific as the lmt_queue is.  Also it improves the queue to use
GENERIC_Q and the 'dev.next' field so it does not need to malloc
additional space.

Signed-off-by: Wesley Bland <wbland@anl.gov>