Return request from IRECV even if failure
We will now return a request handle from MPI_IRECV even if there is a
failure. The reason for this is because the ULFM spec says that even if
the function returns MPIX_ERR_PROC_FAILED_PENDING, it still should
provide a valid request that can be completed later.

This doesn't cause a problem for other situations because the value of
the request is undefined in that scenario so it's fine for it to be
garbage.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>