Move definition of global window counters to nemesis / sock.
num_active_issued_win and num_passive_win are counters of
windows in active ISSUED mode and in passive mode.
It is modified in CH3 and is used in progress engine of
nemesis / sock to skip windows that do not need to make
progress on. Here we define them in mpidi_ch3_pre.h in
nemesis / sock so that they can be exposed to upper layers.

Signed-off-by: Min Si <msi@il.is.s.u-tokyo.ac.jp>