Adds checking VC state in MPID_Send
MPID_Send should first check the state of a VC to see if it is valid
before creating a group, request or communicator.

In the case of fault tolerance, if VC has already been revoked or marked
as terminated (e.g., in test/mpi/ft/senddead). The send operation
evolved should exit without creating any memory objects of request,
group or communicator.

Signed-off-by: Wesley Bland <wbland@anl.gov>