portals4: better dtype mismatch detection
The previous code only detected a datatype mismatch when the message
was copied out of the unexpected queue. Now it will throw an error
in both cases.

We also set the error in the status object to match the default ch3
behavior. This fixed an issue where the request would not be freed
and cause extra debugging output at MPI_Finalize.

Signed-off-by: Antonio Pena Monferrer <apenya@mcs.anl.gov>