Clean up event stashing.
Now, when we pop an event, we queue up the buddy event (e.g., ACK for
SEND) to return next.  This way, we don't need to search for the event
everytime.  Since we know that there'll be at most one such pending
event, we maintain a single event structure for this.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>