fixup test for abort with fault tolerance
MPICH now behaves correctly for this test. There is no reason for it to output
" No errors", since the only thing we are testing for is that it does not
timeout. We also use a non-zero error code in MPI_Abort to fit the requirements
of the test runner. Closes #1537

Signed-off-by: Pavan Balaji <balaji@anl.gov>