Don't free a request if it still pending
If we had a failure that caused a request to be pending, we were freeing
the request before calling the error handler. That caused segfaults. Now
we switch the ordering of the two to avoid that.

This also moves the assignment of the status_ptr to be a little earlier
to avoid another segfault.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>