romio gpfs: select correct read buffer
ROMIO GPFSMPIO_P2PCONTIG threaded read needs to toggle first read buffer

When using both the GPFSMPIO_P2PCONTIG and GPFSMPIO_PTHREADIO
optimizations there was a correctness bug when reading where for the
first round the read buffer did not toggle to the two-phase buffer for
the pthread reader, resulting in diseminating the data from the wrong
buffer.  The fix is to do the toggle after the first read.

Signed-off-by: Paul Coffman <pkcoff@us.ibm.com>
Signed-off-by: Rob Latham <robl@mcs.anl.gov>