Strip out pending ANY_SOURCE request handling
The existing way that we handle non-blocking requests involving wildcard
receive operations is incorrect. We're cancelling request operations and
trying to recreate them later. In the meantime, it's messing with
matching and makes it possible (likely?) that some messages that arrive
will never be matched. A new way of handling this is coming next.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>