remove mpl/opa libs from external flags when embedded
We were incorrectly adding the build directories for mpl and opa to
external_ldflags in Makefile.am, causing them to be listed in the
installed libmpi.la libtool file. If a linker does not handle this
potentially non-existant build directory gracefully, it could cause
an issue. Since the mpl and opa libraries are now embedded in libmpi
by default, we simply eliminate the flags unless we are using
pre-built, external libraries. Fixes #2208

Thanks to Markus Geimer for the bug report and suggested solution.

Signed-off-by: Sangmin Seo <sseo@anl.gov>