Fix win size translation in attrlangf90 test.
This test passed a 0 size to win_create which is translated to a
integer(32bit) var by fortran compiler and passed to c mpi_win_create as
an invalid MPI_Aint(64bit) var by fortran binding because prototype
checking is not supported. This test can be failed if mpi_win_create
internally initializes resource related to the value of size (i.e., mxm
maps win buffer in win_init).

This patch fixed this issue by passing a 64bit local variable as size
parameter instead of a constant var 0 in this f90 test.

Signed-off-by: Junchao Zhang <jczhang@mcs.anl.gov>