romio gpfs: select correct read buffer addendum
The original fix for 'romio gpfs: select correct read buffer' was still
missing a critical piece for the last round to use the correct read
buffer, resulting in a correctness issue that was missed by IOR but
still found by the IBM PE test team.  The fix was to correctly toggle
the buffer after the last read.

Signed-off-by: Paul Coffman <pkcoff@us.ibm.com>
Signed-off-by: Rob Latham <robl@mcs.anl.gov>