Add new pkt flags for different LOCK ACKs.
Add new flags for four different kinds of LOCK ACKs:

(1) LOCK_GRANTED: lock is granted on target.
(2) LOCK_QUEUED_DATA_QUEUED: lock is not granted on target,
    but it is safely queued on target. If this lock request
    is sent with an RMA operation, the operation data is also
    safely queued on target.
(3) LOCK_QUEUED_DATA_DISCARDED: lock is not granted on target,
    but it is safely queued on target. If this lock request
    is sent with an RMA operation, the operation data is discarded
    on target due to out of resources.
(4) LOCK_DISCARDED: lock is not granted on target, and it is
    not queued up on target due to out of resources. If this
    lock request is set with an RMA opration, the operation data
    is also discarded on target.

No reviewer.