Refactor MPIC_Wait to return an errflag
The collective helper functions generally have an errflag that is used
when a failure is detected to allow the collective to continue while
also communicating that a failure occurred. That flag is now included
as a parameter for MPIC_Wait.

The rest of this commit is the refactoring necessary in the rest of the
helper functions to support the change.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>