Refactor MPIC functions to use the MPID objects
The MPIC helper functions have been using MPI_Comm and MPI_Request
objects instead of their MPID_* counterparts. This leads to a bunch of
unnecessary conversions back and forth between the two types of objects
and makes the work incompatible with other parts of the codebase
(non-blocking collectives for instance).

This patch converts all of the MPIC_* functions to use MPID_Comm and
MPID_Request and changes all of the collective calls to use them now
too.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>