Convert scheduler functions to use MPIC
The scheduler functions now use MPIC_* functions to handle their
communication instead of directly calling the MPID_* functions. This
helps to simplify code related to error handling and allows the
collectives to complete even if a failure is detected because the error
will be tracked via the errflag inside the request object.

Fixes #2222

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>