Bug-fix: making processes with SHM and without SHM win work corrrectly.
In commit 7d71278, if node_comm is NULL (only self process is on that
node), we call allocate_no_shm() in CH3 to allocate window. If
node_comm is not NULL (more than one process is on the same node), we
call allocate_shm() in Nemesis to allocate SHM window. However,
the exchanged information amount (in MPI_Allgather) is different
in allocate_no_shm() and allocate_shm(), which leads to wrong execution
when both SHM window and non-SHM window exist. This patch fixes this issue.

Signed-off-by: Pavan Balaji <balaji@anl.gov>