Simplify code of issuing RMA packets.
When issuing RMA packets, we do not need to
store target_win_handle in the request on
origin side but only need to store source_win_handle.
Because when the response data is back, we
only needs to use source_win_handle on origin
size. This patch simplifies the code in this way.

Signed-off-by: Pavan Balaji <balaji@anl.gov>