remove _XOPEN_SOURCE in code
lustre, or specifically the header files lustre brings in (quota.h) show
a problem with caddr_t not being defined if XOPEN_SOURCE is set.  We
provided this define so we could make use of pread/pwrite, but instead
we will use our replacement pread/pwrite if one does not exist.

See #1973

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>