use MPIDI_CH3U_Request_complete to complete user request.
In the request handler, we should use MPIDI_CH3U_Request_complete
to complete user request instead of directly setting it
to being completed. This is because when one operation is cut
into several packets, we must wait until all packets
to be completed to set the user request to be completed.

Signed-off-by: Pavan Balaji <balaji@anl.gov>