Reduce size of shm_base_addrs[] from comm_size to node_size.
Given one process, shm_base_addrs[] is used to store the base
addresses (in the address space of this process) of SHM window
on other processes. The original size of it is comm_size. However,
the maximum number of SHM windows that this process can access
to is node_size instead of comm_size, which results in a waste
of memory since most slots in the array is NULL. In this patch
we reduce the size of shm_base_addrs[] from comm_size to node_size.

Signed-off-by: Pavan Balaji <balaji@anl.gov>