Change name from data_size to buf_size.
When the lock is not satisfied, we queue up
the lock request and op data in a lock entry
queue. In the struct of lock entry, we use 'data_size'
to remember the size of buffer for storing the
data. Since the size of buffer is not type_size*count
but might be type_extent*extent, here we change
its name from 'data_size' to 'buf_size'.

Signed-off-by: Pavan Balaji <balaji@anl.gov>