Code-refactoring: make perform_get_acc_in_lock_queue cleaner.
This patch does not change any functionality but just makes the
code structure cleaner.

The original code structure of perform_get_acc_in_lock_queue is
a mess since the code of dealing with IMMED packet type and the
code of dealing with normal packet type are mixed together.
This patch separates these two parts and makes the function looks
cleaner.

Signed-off-by: Pavan Balaji <balaji@anl.gov>