Code refactoring: setting ureq after issuing op in a function.
After (1) issuing an op (no LOCK flag), or (2) issuing an op
(with LOCK flag) and receiving an ACK that LOCK is granted or
queued, we should set the user request (ureq) to be completed.
This patch wraps up the work of setting ureq into a function,
and call that function after (1) and (2) happens.

Signed-off-by: Pavan Balaji <balaji@anl.gov>