Cutting ACC and GACC messages.
In this patch, we define the size of streaming unit the same
as the SRBuf size (256 * 1024 bytes), and cut the ACC/GACC packet
according to this size. The streaming unit always contains
complete basic type data and does not contain partial basic
type data.

Note that we also increment the ref counter of the pointer
to the derived datatype since multiple streaming units within
one RMA operation will refer to it.

Signed-off-by: Pavan Balaji <balaji@anl.gov>