Reset flags for stream unit within one RMA operation.
For all stream units within one RMA operation, we only
needs to piggyback flags for the first operation to the
first stream unit, and piggyback flags for the last
operation to the last stream unit.

Note that for operations piggybacked with LOCK flag, we
should just issue the first stream unit, and wait until
we receive ACK from the target to decide if we continue
to issue the following units, or  re-transmit the first
unit.

Signed-off-by: Pavan Balaji <balaji@anl.gov>