Add counter in op struct to remember number of stream units issued.
Add a counter in op struct to remember number of stream units
that have already been issued. For example, when the first stream
unit piggybacked with LOCK is issued out, we temporarily stop
issuing the following units. After the origin receives the ACK
from the target, it can continue to issue the following units.
This counter helps avoid issuing the first unit again.

Signed-off-by: Pavan Balaji <balaji@anl.gov>