Add a function hook to initialize window attributes in channel layer.
There are some window attributes in the channel layer that
needs to be initialized during window creation. In this
patch, we first add a win_hooks table that contains pointers
to the channel's implementation of the function hooks. Secondly,
we add a function hook 'win_init' to allow the channel layer to
initialize its own attributes. The hook is called from the
CH3 win_init function.

Signed-off-by: Pavan Balaji <balaji@anl.gov>