Allow the channel layer to implement Win_gather_info function.
In this patch, we first add a function pointer of Win_gather_info
in CH3 to allow different channel layers to implement their own
version of Win_gather_info function. The function pointer is
initially set to the default implementation in CH3 layer. If the
channel layer provides an implementation of Win_gather_info, it
will override the function pointer.

Secondly, we provide an implementation of Win_gather_info in the
Nemesis layer. In this implementation, we allocate basic_info_table[]
in the SHM region, so that processes on the same node can share the
same base_info_table[].

Signed-off-by: Pavan Balaji <balaji@anl.gov>