Use function hook instead of function pointer for win_free.
The original implementation of win_free is not correct. The
problem is described as follows:

It uses a function pointer which is initially set to the CH3
implementation, and can be overridden by the channel layer if
the channel provides an specific implementation.  In the CH3
win_free implementation, it first checks if all RMA
communication is finished and epoch states is reset, then
performs a global barrier, then frees the window resources
that are allocated in CH3, and finally returns. In the Nemesis
win_free implementation, it directly frees the window resources
that are allocated in Nemesis, and calls the CH3 win_free at last.
This makes no sense because we free the window resources before
checking if the RMA communication is completed.

To fix this issue, we add a function hook for channel layer
to free its own resources, the the function hook is called from
the CH3 win_free.

Signed-off-by: Pavan Balaji <balaji@anl.gov>