Do not reference a CH3 variable in ROMIO glue.
Although MPIDI_CH3I_progress_blocked is a variable only used in CH3,
it was referenced in the ROMIO glue code. This caused a build problem
when pamid is used as a device. This patch removed the reference to
MPIDI_CH3I_progress_blocked, but it degrades the efficiency of
MPIR_Ext_cs_yield_allfunc_if_progress_blocked() since we do not have
a way to check if the progress engine is blocked for now (related to
ticket #2202). For a better solution for ticket #2202, we need to
fix a wait function of the extended generalized request.

Fixes #2242

Signed-off-by: Rob Latham <robl@mcs.anl.gov>