Extend timeouts for putfence1/getfence1
These tests take a longer time on OSX due to the poor performance of
malloc/calloc on OSX platforms. The extended timeout gives them a chance
to complete.

This could be better fixed in the future by only allocating the memory
for the window once instead of over and over (~5000 times).

No reviewer