portals4: fix incorrect usage of max_eqs limit
The max_eqs limit returned at PtlNIInit time is actually a limit on the
number of event queues a particular interface can support. It is not the
number of events a single queue can hold, which is how we were using it.

Since there is no way to query for the max events a queue can hold, we
simply request a conservative 32K. Thanks to Sayantan Sur for pointing out
our error.

Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>