gforker: implement PMI_Abort handler
Since [b1e89abf], PMI_Abort sends an abort command to the PMI server
and relies on it to cleanup the application. In the case of gforker,
nothing was done with this command, leaving the PMI client hanging
waiting for a response. Gforker will now exit with the requested abort
code upon receiving the command. Fixes #2249

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>