Fix unsafe datatype release in several RMA tests
The datatypes shouldn't be released until we make sure that there
are no more remote operations using that datatype. I've changed several
tests to release the datatype after a barrier. To avoid introducing
a barrier in every iteration, and aiming to stress out a little more,
I've restructured the tests so that the datatypes are not created and
freed every iteration.

This was causing intermittent segfaults mainly with async enabled.

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>