Simplify the bcast test.
The current number of combinations we are checking are too many,
causing the test to take too long on some platforms.  This patch
simplifies the test, so we build two versions of the test.  In the
first version, we run only on COMM_WORLD but go through all datatypes.
In the second version, we run on all communicators, but go through
only a small subset of datatypes.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>