Fixed the Fortran common symbol issue on Mac.
The linker on Darwin does not allow common symbols, thus libtool adds
the -fno-common option by default for shared libraries. However, the
common symbols defined in different shared libraries and object files
still can not be treated as the same symbol.
For example:
with gfortran, the same common block in the shared libraries and the
object files will have different memory locations separately;
with ifort, the same common block in different shared libraries will get
the same memory location but still get a different location in the
object file.

The -Wl,-commons,use_dylibs option asks linker to check dylibs for
definitions and use them to replace tentative definitions(commons) from
object files, thus it solves the issue of the common symbol mismatch
between the object file and the dylibs (i.e., by setting the address of
a common symbol to the place located in the first dylib that is linked
with the object file and contains this symbol). It needs to be added
only in the linking stage for the final executable file.

The -flat-namespace option allows linker to unify the same common
symbols in different dylibs. It needs to be added in linking stage for
both the shared library and the final executable file.
(see man ld for their definition)

Although gfortran works fine by only adding -flat-namespace, and ifort
works by only adding -Wl,-commons,use_dylibs, we should add both options
here as a generic solution to make sure everything safe.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>