portals4: fix anysource_matched
This fix, along with a pending patch to the Portal4 reference implementation,
should make anysource_matched a more reliable operation for multithreaded apps. We
were seeing a race condition where an ME would unlink successfully, but an
event matching it would still arrive in the queue. CH3 can now reliably search
the netmod queue for matched MPI_ANY_SOURCE requests.

The reason that we no longer assert that an MPI_ANY_SOURCE request was removed
from the CH3 queue is that FDP (find and dequeue posted) operations will remove
the request from the queue, if it is known to be already matched by the netmod,
even if it has not yet completed.

Fixes #2199

Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>