Reenable a bunch of tests that were disabled so far.
Over time, we have disabled a number of tests.  Some of these were
ones that we had tickets for and some were just simply commented out
before we had the xfail mechanism.  This patch is semi-temporary to
see where we stand with these changes and to try to see how far we can
get in fixing them.

This patch does not reenable the FT failures or the performance test
failures.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>