Set size of IMMED data in RMA packets to 8 bytes.
Originally the size of IMMED data in RMA packets is 16 bytes
which makes the size of CH3 packet be 56 bytes. Here we reduce
the size of IMMED data in RMA packets to 8 bytes, so that the
size of CH3 packet is reduced to 48 bytes, the same with
mpich-3.1.4 (the old RMA infrastructure).

Signed-off-by: Min Si <msi@il.is.s.u-tokyo.ac.jp>
Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>