Cleanup threaded progress.
The nemesis progress engine was written in a way so that if one thread
is inside a progress engine, other threads cannot enter the receive
progress.  They can enter the send progress in some cases.  There
doesn't seem to be a good reason for this behavior.  This patch
combines this so threads would simply return for nonblocking
operations and wait for a signal before entering the progress engine
for blocking operations.

Signed-off-by: Halim Amer <aamer@anl.gov>